requirements.txt
requests==2.13.0
google-cloud-bigquery==0.23.0
